requirements.txt
python-dotenv
jsonschema>=4.23.0
pytest>=8.3.0
pyyaml>=6.0.2